Authenticate the Google Data API
The README is not correct in this commit, and I think I will update
it once everything is working and I know what is needed. If you look
at photo_hopper.py, you'll see that there are two files, creds.dat
and client_secrets.json, that are needed but not included in the
repository. These will be explained in the README when I get around
to it.