Refactor module and introduce script flow
No longer does photo_hopper need to store and credentials whatsoever!
All information is input via the command line, and API tokens are
temporary tokens that can be generated online. This increases security,
portability, and ease of use, at least in my opinion. To polish off
this script, I am going to try to port it to a Docker container, so
no dependencies (other than Docker itself) will have to be installed.