Refactor methods into an object oriented design
I want to be able to hop photos in either direction,
so I realized it would make a lot more sense to have a class design
that allows me to add in new services fairly easily. This commit
is entirely a refactor of hopping from Facebook to Google Photos,
and an update of the Dockerfile to add in the new module.